Create reusable buttons. Add new tag editor modal.
