+++ frontend/src/component/SettingsLinkEditor.tsx
@@ -9,7 +9,6 @@ import React, {
 } from 'react';
 import Dialog from '@mui/material/Dialog';
 import {
-    Button,
     Checkbox,
     DialogActions,
     DialogContent,
@@ -23,16 +22,14 @@ import {
     TextField
 } from '@mui/material';
 import Stack from '@mui/material/Stack';
-import CloseIcon from '@mui/icons-material/Close';
-import CheckIcon from '@mui/icons-material/Check';
 import { inputValidator } from './SettingsEditor';
 import MenuItem from '@mui/material/MenuItem';
 import Typography from '@mui/material/Typography';
 import { Trans, useTranslation } from 'react-i18next';
 import NiceModal, { muiDialog, useModal } from '@ebay/nice-modal-react';
 import cloneDeep from 'lodash/cloneDeep';
-import RestartAltIcon from '@mui/icons-material/RestartAlt';
 import { logError } from '../util';
+import { CancelButton, ResetButton, SaveButton } from './Buttons';
 
 interface SettingsLinkProps {
     link: Link;
@@ -158,30 +155,9 @@ export const SettingsLinkEditor = NiceModal.create<SettingsLinkProps>(
                 </DialogContent>
 
                 <DialogActions>
-                    <Button
-                        onClick={modal.hide}
-                        startIcon={<CloseIcon />}
-                        color={'error'}
-                        variant={'contained'}
-                    >
-                        <Trans i18nKey={'button.cancel'} />
-                    </Button>
-                    <Button
-                        onClick={handleReset}
-                        startIcon={<RestartAltIcon />}
-                        color={'warning'}
-                        variant={'contained'}
-                    >
-                        <Trans i18nKey={'button.reset'} />
-                    </Button>
-                    <Button
-                        onClick={handleSave}
-                        startIcon={<CheckIcon />}
-                        color={'success'}
-                        variant={'contained'}
-                    >
-                        <Trans i18nKey={'button.ok'} />
-                    </Button>
+                    <CancelButton />
+                    <ResetButton onClick={handleReset} />
+                    <SaveButton onClick={handleSave} />
                 </DialogActions>
             </Dialog>
         );
